+++ examples/airflow_pool/main.tf
@@ -15,7 +15,9 @@
  */
 
 module "simple-composer-environment" {
-  source                           = "../../modules/create_environment_v1"
+  source  = "terraform-google-modules/composer/google//modules/create_environment_v1"
+  version = "~> 4.0"
+
   project_id                       = var.project_id
   composer_env_name                = var.composer_env_name
   region                           = var.region
@@ -27,11 +29,14 @@ module "simple-composer-environment" {
   service_ip_allocation_range_name = var.service_ip_allocation_range_name
   node_count                       = 3
   machine_type                     = "n1-standard-1"
+  image_version                    = "composer-1.20.12-airflow-1.10.15"
 }
 
 # Making the k8s master globally available is only to make the integration testing portable and should be removed
 module "master-authorized-networks" {
-  source      = "../../modules/master_authorized_networks"
+  source  = "terraform-google-modules/composer/google//modules/master_authorized_networks"
+  version = "~> 4.0"
+
   project_id  = var.project_id
   zone        = var.zone
   gke_cluster = module.simple-composer-environment.gke_cluster
@@ -42,7 +47,9 @@ module "master-authorized-networks" {
 
 # Pools can be defined externally if you wish
 module "pool-1" {
-  source            = "../../modules/airflow_pool"
+  source  = "terraform-google-modules/composer/google//modules/airflow_pool"
+  version = "~> 4.0"
+
   project_id        = var.project_id
   composer_env_name = module.simple-composer-environment.composer_env_name
   region            = var.region