+++ examples/composer_v2_sharedvpc_prereq/composer.tf
@@ -13,7 +13,9 @@
 # limitations under the License.
 
 module "composer_net" {
-  source                                 = "../../modules/composer_net"
+  source  = "terraform-google-modules/composer/google//modules/composer_net"
+  version = "~> 4.0"
+
   service_project_id                     = var.service_project_id
   network_project_id                     = var.project_id
   composer_env_name                      = "san-composer-2"
@@ -22,7 +24,6 @@ module "composer_net" {
   subnetwork                             = "composer-subnet"
   cloud_composer_network_ipv4_cidr_block = "192.168.192.0/24"
   master_ipv4_cidr                       = "192.168.193.0/28"
-  cloud_sql_ipv4_cidr                    = "192.168.0.0/17"
   gke_subnet_ip_range                    = ["10.100.232.0/27"]
   gke_pods_services_ip_ranges            = ["10.1.0.0/16", "10.4.0.0/16", "10.10.10.0/24", "10.10.14.0/24"]
 }
@@ -30,7 +31,9 @@ module "composer_env" {
   depends_on = [
     module.composer_net
   ]
-  source                                 = "../../modules/create_environment_v2"
+  source  = "terraform-google-modules/composer/google//modules/create_environment_v2"
+  version = "~> 4.0"
+
   project_id                             = var.service_project_id
   network_project_id                     = var.project_id
   composer_env_name                      = "san-composer-2"