+++ modules/create_environment_v2/versions.tf
@@ -15,26 +15,26 @@
  */
 
 terraform {
-  required_version = ">= 0.13"
+  required_version = ">= 1.3"
   required_providers {
 
     google = {
       source  = "hashicorp/google"
-      version = ">= 4.8.0, < 5.0"
+      version = ">= 5.14, < 6"
     }
 
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = ">= 4.8.0, < 5.0"
+      version = ">= 5.14, < 6"
     }
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-composer:create_environment_v2/v3.4.0"
+    module_name = "blueprints/terraform/terraform-google-composer:create_environment_v2/v5.0.0"
   }
 
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-composer:create_environment_v2/v3.4.0"
+    module_name = "blueprints/terraform/terraform-google-composer:create_environment_v2/v5.0.0"
   }
 
 }