+++ CHANGELOG.md
@@ -8,6 +8,43 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 This changelog is generated automatically based on [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
 
+## [5.0.0](https://github.com/terraform-google-modules/terraform-google-composer/compare/v4.1.0...v5.0.0) (2024-03-07)
+
+
+### ⚠ BREAKING CHANGES
+
+* **TPG>=5.14:** add data_retention_config block ([#121](https://github.com/terraform-google-modules/terraform-google-composer/issues/121))
+
+### Features
+
+* add cmek support in composer_env_v2 ([#113](https://github.com/terraform-google-modules/terraform-google-composer/issues/113)) ([0fe2605](https://github.com/terraform-google-modules/terraform-google-composer/commit/0fe2605df22e658212e5a3b9048939ecf2837bcc))
+* **TPG>=5.14:** add data_retention_config block ([#121](https://github.com/terraform-google-modules/terraform-google-composer/issues/121)) ([ca0dfcc](https://github.com/terraform-google-modules/terraform-google-composer/commit/ca0dfcca64f3e3103a7ece9f2f3bffa43e2bfa95))
+
+## [4.1.0](https://github.com/terraform-google-modules/terraform-google-composer/compare/v4.0.1...v4.1.0) (2023-12-04)
+
+
+### Features
+
+* add web_server_network_access_control ([#102](https://github.com/terraform-google-modules/terraform-google-composer/issues/102)) ([6064816](https://github.com/terraform-google-modules/terraform-google-composer/commit/60648169ee59b28c4c300d82c9b80db602b0670b))
+
+## [4.0.1](https://github.com/terraform-google-modules/terraform-google-composer/compare/v4.0.0...v4.0.1) (2023-11-22)
+
+
+### Bug Fixes
+
+* duplicate maintenance_window ([#100](https://github.com/terraform-google-modules/terraform-google-composer/issues/100)) ([82e3c04](https://github.com/terraform-google-modules/terraform-google-composer/commit/82e3c041d90cc6e8173419d40c462bd525a1d892))
+
+## [4.0.0](https://github.com/terraform-google-modules/terraform-google-composer/compare/v3.4.0...v4.0.0) (2023-11-16)
+
+
+### ⚠ BREAKING CHANGES
+
+* **TPG >=5.3:** add new features. update provider requirements ([#98](https://github.com/terraform-google-modules/terraform-google-composer/issues/98))
+
+### Features
+
+* **TPG >=5.3:** add new features. update provider requirements ([#98](https://github.com/terraform-google-modules/terraform-google-composer/issues/98)) ([ea834f7](https://github.com/terraform-google-modules/terraform-google-composer/commit/ea834f73eb3f55f6c1a6f2c1c0540852c0836abc))
+
 ## [3.4.0](https://github.com/terraform-google-modules/terraform-google-composer/compare/v3.3.0...v3.4.0) (2022-11-28)